pyproject.toml
[tool.poetry]
name = "rickandmorty integeration"
version = "0.1.0"
description = ""
authors = ["Alex <124603897+OleksandrBrusyltsev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
